requirements.txt
streamlit==1.40.0
agno==2.4.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.2.0
sqlalchemy==2.0.25
openai>=1.0.0
